before after methods added and insert fixed
