+++ README.md
@@ -10,4 +10,4 @@
 challenge 01 - [Array Reverse](challenges/arrayReverse/array-shift.js)
 challenge 02 - [Array Shift](challenges/arrayShift/array-shift.js)
 challenge 03 - [Array Binary Search](challenges/arrayBinarySearch/array-binary-search.js)
-challenge 05 - [Linked lists](challenges/linkedList/linked-list.js)
+challenge 05 - [Linked lists](challenges/Data-Structures/linkedList/linked-list.js)
